+++ pkg/stdcopy/MAINTAINERS
@@ -1,0 +1,1 @@
+Cristian Staretu <cristian.staretu@gmail.com> (@unclejack)
