+++ pkg/stdcopy/stdcopy.go
@@ -1,4 +1,4 @@
-package utils
+package stdcopy
 
 import (
 	"encoding/binary"
